chore: Rename namespace.
Fixes #50
